Moved `der_sort()` back.
